content/svg/content/test/test_pointAtLength.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=643419
-->
<head>
  <title>Test getPointAtLength</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function run()
{
  var p1 = document.getElementById("p1");
  var point = p1.getPointAtLength(200);
  is(point.x, 200);
  is(point.y, 50);

  // set the pathLength to twice its actual length
  p1.setAttribute("pathLength", "800");
  var point = p1.getPointAtLength(200);
  is(point.x, 100);
  is(point.y, 50);

  SimpleTest.finish();
}

window.addEventListener("load", run, false);
]]>
</script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=643419">Mozilla Bug 643419</a>
<p id="display"></p>
<div id="content">

  <svg xmlns="http://www.w3.org/2000/svg" width="750">
    <defs>
      <path id="p1" d="M 0 50 h 400"/>
    </defs>
  </svg>

</div>
<pre id="test">
</pre>
</body>
</html>
